Connect Redis as cache
