+++ data_search/src/main/java/com/example/datasearchservice/configuration/RedisCacheConfig.java
@@ -1,28 +1,64 @@
 package com.example.datasearchservice.configuration;
 
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.cache.annotation.EnableCaching;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.cache.RedisCacheConfiguration;
 import org.springframework.data.redis.cache.RedisCacheManager;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
+import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
 
 import java.time.Duration;
 import java.util.Collections;
 
 @Configuration
+@EnableCaching
 public class RedisCacheConfig {
+//    @Bean
+//    public RedisCacheManager cacheManager(RedisConnectionFactory connectionFactory) {
+//        return RedisCacheManager.builder(connectionFactory)
+//                .cacheDefaults(RedisCacheConfiguration.defaultCacheConfig())
+//                .transactionAware()
+//                .withInitialCacheConfigurations(Collections.singletonMap("resources", cacheConfiguration()))
+//                .build();
+//    }
+//    private RedisCacheConfiguration cacheConfiguration() {
+//        return RedisCacheConfiguration.defaultCacheConfig()
+//                .entryTtl(Duration.ofSeconds(120))
+//                .disableCachingNullValues();
+//    }
+
+    @Value("${SPRING_REDIS_HOST:localhost}")
+    private String redisHost;
+
+    @Value("${SPRING_REDIS_PORT:6379}")
+    private int redisPort;
+
     @Bean
-    public RedisCacheManager cacheManager(RedisConnectionFactory connectionFactory) {
-        return RedisCacheManager.builder(connectionFactory)
-                .cacheDefaults(RedisCacheConfiguration.defaultCacheConfig())
-                .transactionAware()
-                .withInitialCacheConfigurations(Collections.singletonMap("resources", cacheConfiguration()))
+    public RedisConnectionFactory redisConnectionFactory() {
+        return new LettuceConnectionFactory(redisHost, redisPort);
+    }
+
+    @Bean
+    public RedisCacheConfiguration cacheConfiguration() {
+        return RedisCacheConfiguration.defaultCacheConfig()
+                .entryTtl(Duration.ofMinutes(10)) // Cache expiry time
+                .disableCachingNullValues();
+    }
+
+    @Bean
+    public RedisCacheManager cacheManager(RedisConnectionFactory redisConnectionFactory) {
+        return RedisCacheManager.builder(redisConnectionFactory)
+                .cacheDefaults(cacheConfiguration())
                 .build();
     }
-    private RedisCacheConfiguration cacheConfiguration() {
-        return RedisCacheConfiguration.defaultCacheConfig()
-                .entryTtl(Duration.ofSeconds(120))
-                .disableCachingNullValues();
+
+    @Bean
+    public RedisTemplate<Object, Object> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
+        RedisTemplate<Object, Object> template = new RedisTemplate<>();
+        template.setConnectionFactory(redisConnectionFactory);
+        return template;
     }
 }
