+++ data_search/src/main/java/com/example/datasearchservice/service/SearchService.java
@@ -1,7 +1,9 @@
 package com.example.datasearchservice.service;
 
 import com.example.datasearchservice.entity.Resource;
+import com.example.datasearchservice.models.AdvancedSearchDTO;
 import com.example.datasearchservice.repository.ResourceRepository;
+import org.springframework.cache.annotation.Cacheable;
 import org.springframework.stereotype.Service;
 
 
@@ -20,8 +22,12 @@ public class SearchService {
     public List<Resource> searchByTags(List<String> tags) {
         return repository.findByTagsContaining(tags);
     }
-    public List<Resource> generalSearch(String content, List<String> tags) {
-        return repository.findByTagsOrContent(tags, content);
+    @Cacheable(value = "searchCache", key = "#searchPhrase")
+    public List<Resource> generalSearch(String searchPhrase) {
+        AdvancedSearchDTO searchDTO = new AdvancedSearchDTO();
+        searchDTO.content = searchPhrase;
+        searchDTO.tags = List.of(searchPhrase.split(" "));
+        return repository.findByTagsOrContent(searchDTO.tags, searchDTO.content);
     }
 
 }
