+++ data_search/src/main/resources/application.properties
@@ -1,11 +1,13 @@
 spring.application.name=dataSearchService
-server.servlet.contextPath=/api
+server.servlet.contextPath=/api/v1
 
 spring.docker.compose.enabled=false
+spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration
 
 spring.data.redis.host=localhost
 spring.data.redis.port=6379
-spring.autoconfigure.exclude=org.springframework.boot.autoconfigure.security.servlet.SecurityAutoConfiguration
+spring.cache.type=redis
+
 
 spring.data.elasticsearch.repositories.enabled=true
 spring.elasticsearch.uris=http://localhost:9200
@@ -16,4 +18,3 @@ logging.level.org.springframework.data.redis=DEBUG
 logging.level.org.springframework.cache=DEBUG
 logging.level.org.springframework.cache.interceptor=DEBUG
 
-spring.cache.type=redis
